requirements.txt
#Основной фреймворк для бота
python-telegram-bot[job-queue]~=20.7
#Для работы с переменным окружением
python-dotenv~=1.0.0
#Ассинхронная работа с SQLite
aiosqlite~=0.19.0
#Для работы с изображениями
Pillow~=10.0.0
